+++ dll/ime/msctfime/msctfime.h
@@ -25,8 +25,7 @@
 #include <cicero/cicbase.h>
 #include <cicero/cicarray.h>
 #include <cicero/cicpath.h>
-#include <cicero/osinfo.h>
-#include <cicero/imclock.h>
+#include <cicero/cicimc.h>
 
 class CicInputContext;
